backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.30.1
pdfplumber==0.11.0
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.3.0
beautifulsoup4==4.12.3
requests==2.32.3
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
aiofiles==23.2.1
lxml==5.2.2
